Admin_bot/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
imgkit==1.2.3
magic-filter==1.0.12
multidict==6.0.4
psycopg2-binary==2.9.9
pydantic==1.10.13
six==1.16.0
typing_extensions==4.9.0
yarl==1.9.4